+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 * Support of adding any widget to VerticalMenuForAndroidTV, it won't get focused
 
+* Added example code in README.md
+
 ## [0.0.3] - 1 Sep 2019
 
 * Bug fix, VerticalMenuItem needs to extend from StatelessWidget